refactor(handlers): migrate mvp + kept-module handlers to keyboards_v2
