+++ projects/polymarket/crusaderbot/bot/handlers/mvp/portfolio.py
@@ -7,7 +7,7 @@ from telegram import Update
 from telegram.ext import Application, CallbackQueryHandler, ContextTypes
 
 from ... import messages_mvp as mvp
-from ...keyboards.mvp import portfolio as kb
+from ...keyboards_v2.mvp import portfolio as kb
 from . import _users
 from ._send import callback_parts, send_or_edit
